Backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# File handling
python-multipart==0.0.12

# PDF conversion and processing
convertapi>=1.4.0  # For technician report conversion
pdfplumber>=0.10.0  # For daily and rework report conversion

# Data processing
pandas==2.2.0
numpy>=1.26.0,<2.0.0
openpyxl==3.1.2  # Excel file support for XLSX output

# Database
mysql-connector-python>=8.0.33

# Additional pandas dependencies
python-dateutil>=2.8.2
pytz>=2023.3  # Timezone support for pandas

# Environment variables
python-dotenv>=1.0.0  # For loading .env configuration

# Type hints and compatibility
typing-extensions>=4.0.0

# REQUIREMENTS:
# 1. Python 3.8+ is required
# 2. ConvertAPI account and API token (set in .env file as CONVERTAPI_TOKEN)
# 3. Create .env file in Backend directory with:
#    CONVERTAPI_TOKEN=your_api_token_here
